backend/requirements.txt
fastapi>=0.100
uvicorn>=0.20
sqlalchemy>=2.0
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
httpx>=0.24
python-multipart>=0.0.5
